SR-Rerouted handles connections with SRH
There are new unit tests as well

Key detail: The packet inside the ICMP is changed so that its destination
is always the final destination of the packet. This might not be the
case because of Segment Routing and the linux kernel cannot retrieve the
connection targeted by the ICMP because of that.